app/requirement.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.2
Django==2.1.2
django-tinymce==3.6.1
idna==2.7
more-itertools==4.3.0
Pillow==5.3.0
pluggy==0.7.1
py==1.7.0
pytest==3.10.0
pytest-cov==2.6.0
pytest-django==3.4.3
python-dateutil==2.7.5
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.23
